requirements.txt
requests>=2.31.0
urllib3>=2.0.0
python-telegram-bot>=21.4
